Complete SchoolManagement Backend with controllers, services, models, repositories, DTOs, and security

+++ src/main/java/com/example/enicarthage/demo_proj_A/Controllers/TeacherController.java
@@ -27,7 +27,7 @@ public class TeacherController {
 
     // Récupérer toutes les notes d’un professeur
     @GetMapping("/{professeurId}/notes")
-    public List<Note> getNotesByProf(@PathVariable Long professeurId) {
+    public List<Note> getNotesByProf(@PathVariable int professeurId) {
         return teacherService.getNotesByProfesseurId(professeurId);
     }
